IS-1981: Refactor moter folder into /sider folder
