+++ src/components/dialogmote/DialogmoteFerdigstilteReferatPanel.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement } from "react";
 import { DialogmoteDTO } from "@/data/dialogmote/types/dialogmoteTypes";
-import { DialogmotePanel } from "@/components/mote/components/DialogmotePanel";
+import { DialogmotePanel } from "@/sider/mote/components/DialogmotePanel";
 import { BlueDocumentImage } from "../../../img/ImageComponents";
 import dayjs from "dayjs";
 import {
