+++ test/containers/MotelandingssideContainerTest.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { expect } from "chai";
-import Motelandingsside from "../../src/components/mote/components/Motelandingsside";
+import Motelandingsside from "@/sider/mote/components/Motelandingsside";
 import { QueryClientProvider } from "@tanstack/react-query";
 import { dialogmoterQueryKeys } from "@/data/dialogmote/dialogmoteQueryHooks";
 import {
