+++ test/dialogmote/DialogmoteMoteStatusPanelTest.tsx
@@ -4,7 +4,7 @@ import {
 } from "@/data/dialogmote/types/dialogmoteTypes";
 import { render, screen } from "@testing-library/react";
 import React from "react";
-import { DialogmoteMoteStatusPanel } from "@/components/mote/components/innkalling/DialogmoteMoteStatusPanel";
+import { DialogmoteMoteStatusPanel } from "@/sider/mote/components/innkalling/DialogmoteMoteStatusPanel";
 import { dialogmote, dialogmoteMedMellomlagretReferat } from "./testData";
 import { QueryClientProvider } from "@tanstack/react-query";
 import { MemoryRouter } from "react-router-dom";
